Splitting up pipeline tests into multiple files
